Support ArgumentValidators on the method that apply to the first argument + [Breaking] Rename command.types to command.validators to match its contents
