+++ base/src/main/java/uk/co/thinkofdeath/command/ArgumentNode.java
@@ -17,7 +17,7 @@
 package uk.co.thinkofdeath.command;
 
 import uk.co.thinkofdeath.command.parsers.ArgumentParser;
-import uk.co.thinkofdeath.command.types.ArgumentValidator;
+import uk.co.thinkofdeath.command.validators.ArgumentValidator;
 
 import java.util.Arrays;
