+++ base/src/main/java/uk/co/thinkofdeath/command/validators/TypeHandler.java
@@ -14,42 +14,17 @@
  * limitations under the License.
  */
 
-package uk.co.thinkofdeath.command.types;
-
-import uk.co.thinkofdeath.command.CommandError;
+package uk.co.thinkofdeath.command.validators;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@TypeHandler(
-        value = MaxLengthHandler.class,
-        clazz = String.class
-)
-@Target(ElementType.PARAMETER)
+@Target(ElementType.ANNOTATION_TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-/**
- * Requires the passed string to have a length less than
- * or equal to the value
- */
-public @interface MaxLength {
-    int value();
-}
+public @interface TypeHandler {
+    Class<? extends ArgumentValidator> value();
 
-class MaxLengthHandler implements ArgumentValidator<String> {
-
-    private final int max;
-
-    MaxLengthHandler(MaxLength maxLength) {
-        max = maxLength.value();
-    }
-
-    @Override
-    public CommandError validate(String argStr, String argument) {
-        if (argument.length() > max) {
-            return new CommandError(3, "validator.maxlength", argument, max);
-        }
-        return null;
-    }
+    Class<?> clazz();
 }
